+++ skbio/tree/tests/test_tree.py
@@ -1491,45 +1491,59 @@ class TreeTests(TestCase):
             self.assertFalse(hasattr(node, "_subset"))
 
         # include singles
-        self.assertSetEqual(t.subsets(include_singles=True), frozenset(
+        self.assertSetEqual(t.subsets(include_single=True), frozenset(
             [frozenset("ab"), frozenset("cd"), frozenset("a"), frozenset("b"),
              frozenset("c"), frozenset("d")]))
 
-    def test_bipartition(self):
+        # map to length
+        t = TreeNode.read(["((a:1,b:2):1,c:4,((d:4,e:5):2,f:6):1);"])
+        self.assertDictEqual(t.subsets(map_to_length=True), {
+            frozenset("ab"): 1, frozenset("de"): 2, frozenset("def"): 1})
+
+        # merged lengths
+        self.assertDictEqual(t.subsets(within="abde", map_to_length=True), {
+            frozenset("ab"): 1, frozenset("de"): 3})
+
+        # missing length
+        t.children[0].length = None
+        self.assertDictEqual(t.subsets(map_to_length=True), {
+            frozenset("ab"): 0, frozenset("de"): 2, frozenset("def"): 1})
+
+    def test_bipart(self):
         """Return a set of tip names on the smaller side of the branch."""
         t = self.complex_tree
 
         # typical cases (internal node): get smaller side
         # taxa below node (same as subset)
-        self.assertSetEqual(t.find("int1").bipartition(), frozenset("ab"))
-        self.assertSetEqual(t.find("int3").bipartition(), frozenset("cd"))
-        self.assertSetEqual(t.find("int5").bipartition(), frozenset("ef"))
+        self.assertSetEqual(t.find("int1").bipart(), frozenset("ab"))
+        self.assertSetEqual(t.find("int3").bipart(), frozenset("cd"))
+        self.assertSetEqual(t.find("int5").bipart(), frozenset("ef"))
 
         # when clade is larger than half of tree, get remaining taxa
-        self.assertSetEqual(t.find("int4").bipartition(), frozenset("abef"))
+        self.assertSetEqual(t.find("int4").bipart(), frozenset("abef"))
 
         # at root: empty set
-        self.assertSetEqual(t.bipartition(), frozenset())
+        self.assertSetEqual(t.bipart(), frozenset())
 
         # at tip: singleton
-        self.assertSetEqual(t.find("a").bipartition(), frozenset("a"))
+        self.assertSetEqual(t.find("a").bipart(), frozenset("a"))
 
         # when size is equal, get lexicographically smaller side
         t = self.simple_t
-        self.assertSetEqual(t.find("i1").bipartition(), frozenset("ab"))
-        self.assertSetEqual(t.find("i2").bipartition(), frozenset("ab"))
+        self.assertSetEqual(t.find("i1").bipart(), frozenset("ab"))
+        self.assertSetEqual(t.find("i2").bipart(), frozenset("ab"))
 
         # an unrooted tree
         t = TreeNode.read(["((a,(b,c))X,(d,e)Y,f);"])
-        self.assertSetEqual(t.find("X").bipartition(), frozenset("abc"))
-        self.assertSetEqual(t.find("Y").bipartition(), frozenset("de"))
+        self.assertSetEqual(t.find("X").bipart(), frozenset("abc"))
+        self.assertSetEqual(t.find("Y").bipart(), frozenset("de"))
 
-    def test_bipartitions(self):
+    def test_biparts(self):
         """Return all sets of tip names on the smaller side of each branch."""
 
         # an unrooted, bifurcating tree (typical case)
         t = TreeNode.read(["(((a,(b,c)),d),(e,f),g);"])
-        self.assertSetEqual(t.bipartitions(), frozenset({
+        self.assertSetEqual(t.biparts(), frozenset({
             frozenset({"e", "f"}),
             frozenset({"e", "f", "g"}),
             frozenset({"b", "c"}),
@@ -1537,7 +1551,7 @@ class TreeTests(TestCase):
 
         # a rooted tree with polytomy
         t = self.complex_tree
-        self.assertSetEqual(t.bipartitions(), frozenset({
+        self.assertSetEqual(t.biparts(), frozenset({
             frozenset({"e", "f"}),
             frozenset({"c", "d"}),
             frozenset({"a", "b"}),
@@ -1545,33 +1559,47 @@ class TreeTests(TestCase):
             frozenset({"a", "b", "e", "f"})}))
 
         # within given taxa
-        self.assertSetEqual(t.bipartitions(within="abcdef"), frozenset({
+        self.assertSetEqual(t.biparts(within="abcdef"), frozenset({
             frozenset({"c", "d"}),
             frozenset({"a", "b"}),
             frozenset({"e", "f"})}))
 
         self.assertSetEqual(
-            t.bipartitions(within="cbda"),
-            t.bipartitions(within={"a", "b", "c", "d"}))
+            t.biparts(within="cbda"),
+            t.biparts(within={"a", "b", "c", "d"}))
 
         # within a subtree
-        self.assertSetEqual(t.children[0].bipartitions(), frozenset({
+        self.assertSetEqual(t.children[0].biparts(), frozenset({
             frozenset({"c", "d"}),
             frozenset({"a", "b"}),
             frozenset({"w", "z"})}))
 
         # a simple tree with only one bipartition
         t = self.simple_t
-        self.assertSetEqual(t.bipartitions(), frozenset({
+        self.assertSetEqual(t.biparts(), frozenset({
             frozenset({"a", "b"})}))
 
         # include singletons (tips)
-        self.assertSetEqual(t.bipartitions(include_singles=True), frozenset({
+        self.assertSetEqual(t.biparts(include_single=True), frozenset({
             frozenset({"a", "b"}),
             frozenset({"a"}),
             frozenset({"b"}),
             frozenset({"c"}),
             frozenset({"d"})}))
+
+        # map to length ("abc" crosses the root)
+        t = TreeNode.read(["(((a:1,b:2):1,c:3):2,((d:4,e:5):2,f:6):2);"])
+        self.assertDictEqual(t.biparts(map_to_length=True), {
+            frozenset("ab"): 1, frozenset("de"): 2, frozenset("abc"): 4})
+
+        # merged lengths
+        self.assertDictEqual(t.biparts(within="abde", map_to_length=True), {
+            frozenset("ab"): 7})
+
+        # missing length
+        t.children[0].length = None
+        self.assertDictEqual(t.biparts(map_to_length=True), {
+            frozenset("ab"): 1, frozenset("de"): 2, frozenset("abc"): 2})
 
     def test_assign_supports(self):
         """Extract support values of internal nodes."""
@@ -1908,13 +1936,17 @@ class TreeTests(TestCase):
         self.assertEqual(t_dm, exp_t_dm)
 
     def test_compare_rfd(self):
-        """Return the Robinson-Foulds distance."""
+        """Return Robinson-Foulds distance."""
         # original example          
-        t = TreeNode.read(["((H,G),(R,M));"])
+        t1 = TreeNode.read(["((H,G),(R,M));"])
         t2 = TreeNode.read(["(((H,G),R),M);"])
-        self.assertEqual(t.compare_rfd(t2), 2)
-        self.assertEqual(t.compare_rfd(t2), t2.compare_rfd(t))
-        self.assertEqual(t.compare_rfd(t2, proportion=True), 0.5)
+        self.assertEqual(t1.compare_rfd(t2), 2)
+
+        # zero distance to self
+        self.assertEqual(t1.compare_rfd(t2), t2.compare_rfd(t1))
+
+        # return proportion
+        self.assertEqual(t1.compare_rfd(t2, proportion=True), 0.5)
 
         # two conflicting quartets
         t1 = TreeNode.read(["((a,b),(c,d));"])
@@ -1949,6 +1981,42 @@ class TreeTests(TestCase):
         self.assertEqual(t1.compare_rfd(t2), 8)
         self.assertEqual(t1.compare_rfd(t2, rooted=False), 6)
 
+    def test_compare_wrfd(self):
+        """Return weighted Robinson-Foulds distance or variants."""
+        t1 = TreeNode.read(["((a:1,b:2):1,c:4,((d:4,e:5):2,f:6):1);"])
+        t2 = TreeNode.read(["((a:3,(b:2,c:2):1):3,d:8,(e:5,f:6):2);"])
+
+        # weighted RF distance (matches phangorn's wRF.dist and dendropy's
+        # weighted_robinson_foulds_distance)
+        self.assertAlmostEqual(t1.compare_wrfd(t2), 16)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="wrf"), 16)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="cityblock"), 16)
+
+        # KF branch score distance (matches phangorn's KF.dist and dendropy's
+        # euclidean_distance)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="kf"), 6.1644140)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="bs"), 6.1644140)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="euclidean"), 6.1644140)
+
+        # no terminal branches (matches ape's dist.topo with method="score")
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="kf", include_single=False),
+                               3.7416574)
+
+        # force rooted (considers subsets not bipartitions)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, rooted=True), 18)
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="kf", rooted=True), 6.6332496)
+
+        # correlation distance
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="correlation"), 0.3164447)
+
+        # unit correlation distance
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="unitcorr"), 0.1582224)
+
+        # correlation distance is independent of scale
+        for node in t2.traverse(include_self=False):
+            node.length *= 10
+        self.assertAlmostEqual(t1.compare_wrfd(t2, metric="correlation"), 0.3164447)
+
     def test_compare_subsets(self):
         """Return the amount of unshared subsets."""
         t = TreeNode.read(["((H,G),(R,M));"])
@@ -1961,11 +2029,17 @@ class TreeTests(TestCase):
         result = t.compare_subsets(t2)
         self.assertEqual(result, 0.5)
 
+        result = t.compare_subsets(t4)
+        self.assertEqual(result, 0.6)
+
         result = t.compare_subsets(t4, proportion=False)
         self.assertEqual(result, 3)
 
         result = t.compare_subsets(t4, shared_only=True)
         self.assertAlmostEqual(result, 1 / 3)
+
+        result = t.compare_subsets(t4, symmetric=False)
+        self.assertEqual(result, 0.5)
 
         result = t.compare_subsets(self.TreeRoot, exclude_absent_taxa=True)
         self.assertEqual(result, 1)
@@ -1973,25 +2047,34 @@ class TreeTests(TestCase):
         result = t.compare_subsets(self.TreeRoot)
         self.assertEqual(result, 1)
 
-    def test_compare_bipartitions(self):
+        result = TreeNode('x').compare_subsets(TreeNode('y'))
+        self.assertEqual(result, 1)
+
+    def test_compare_biparts(self):
         """Return the amount of unshared bipartitions."""
         t = TreeNode.read(["((H,G),(R,M));"])
         t2 = TreeNode.read(["(((H,G),R),M);"])
         t3 = TreeNode.read(["(((H,R),G),M);"])
 
-        result = t.compare_bipartitions(t)
+        result = t.compare_biparts(t)
         self.assertEqual(result, 0)
 
-        result = t.compare_bipartitions(t2)
+        result = t.compare_biparts(t2)
         self.assertEqual(result, 0)
 
-        result = t.compare_bipartitions(t3)
+        result = t.compare_biparts(t3)
         self.assertEqual(result, 1)
 
-        result = t.compare_bipartitions(t3, proportion=False)
+        result = t.compare_biparts(t3, proportion=False)
         self.assertEqual(result, 2)
 
-        result = t.compare_bipartitions(self.TreeRoot)
+        result = t.compare_biparts(t3, proportion=False, symmetric=False)
+        self.assertEqual(result, 1)
+
+        result = t.compare_biparts(self.TreeRoot)
+        self.assertEqual(result, 1)
+
+        result = TreeNode('x').compare_biparts(TreeNode('y'))
         self.assertEqual(result, 1)
 
     def test_compare_tip_distances(self):